Conflicts:
	views/base.tx
	views/edit.tx
	views/edit_complex.tx
	views/index.tx
	views/list.tx
	views/view.tx
	views/view_graph.tx

+++ lib/GrowthForecast/Data.pm
@@ -17,7 +17,8 @@ sub new {
     my $class = shift;
     my $data_dir = shift;
     my $float_number = shift;
-    bless { data_dir => $data_dir, float_number => $float_number }, $class;
+    my $disable_subtract = shift;
+    bless { data_dir => $data_dir, float_number => $float_number, disable_subtract => $disable_subtract }, $class;
 }
 
 sub number_type {
@@ -52,9 +53,10 @@ CREATE TABLE IF NOT EXISTS graphs (
     sllimit      $number_type NOT NULL DEFAULT 0,
     type         VARCHAR(255) NOT NULL DEFAULT 'AREA',
     stype         VARCHAR(255) NOT NULL DEFAULT 'AREA',
-    meta         TEXT NOT NULL DEFAULT '',
+    meta         TEXT,
     created_at   UNSIGNED INT NOT NULL,
     updated_at   UNSIGNED INT NOT NULL,
+    timestamp    UNSIGNED INT DEFAULT NULL,
     UNIQUE  (service_name, section_name, graph_name)
 )
 EOF
@@ -70,7 +72,8 @@ EOF
         }
         $dbh->commit;
 
-        $dbh->do(<<EOF);
+        unless ( $self->{disable_subtract} ) {
+            $dbh->do(<<EOF);
 CREATE TABLE IF NOT EXISTS prev_graphs (
     graph_id     INT NOT NULL,
     number       $number_type NOT NULL DEFAULT 0,
@@ -80,7 +83,7 @@ CREATE TABLE IF NOT EXISTS prev_graphs (
 )
 EOF
 
-        $dbh->do(<<EOF);
+            $dbh->do(<<EOF);
 CREATE TABLE IF NOT EXISTS prev_short_graphs (
     graph_id     INT NOT NULL,
     number       $number_type NOT NULL DEFAULT 0,
@@ -89,6 +92,7 @@ CREATE TABLE IF NOT EXISTS prev_short_graphs (
     PRIMARY KEY  (graph_id)
 )
 EOF
+        }
 
         $dbh->do(<<EOF);
 CREATE TABLE IF NOT EXISTS complex_graphs (
@@ -99,12 +103,52 @@ CREATE TABLE IF NOT EXISTS complex_graphs (
     number       $number_type NOT NULL DEFAULT 0,
     description  VARCHAR(255) NOT NULL DEFAULT '',
     sort         UNSIGNED INT NOT NULL DEFAULT 0,
-    meta         TEXT NOT NULL DEFAULT '',
+    meta         TEXT,
     created_at   UNSIGNED INT NOT NULL,
     updated_at   UNSIGNED INT NOT NULL,
     UNIQUE  (service_name, section_name, graph_name)
 )
 EOF
+
+        $dbh->do(<<EOF);
+CREATE TABLE IF NOT EXISTS vrules (
+    id           INTEGER NOT NULL PRIMARY KEY,
+    graph_path   VARCHAR(255) NOT NULL,
+    time         INT UNSIGNED NOT NULL,
+    color        VARCHAR(255) NOT NULL DEFAULT '#FF0000',
+    description  TEXT,
+    dashes       VARCHAR(255) NOT NULL DEFAULT ''
+)
+EOF
+        $dbh->do(<<EOF);
+CREATE INDEX IF NOT EXISTS time_graph_path on vrules (time, graph_path)
+EOF
+
+        {
+            $dbh->begin_work;
+            my $columns = $dbh->select_all(q{PRAGMA table_info("vrules")});
+            my %graphs_columns;
+            $graphs_columns{$_->{name}} = 1 for @$columns;
+            if ( ! exists $graphs_columns{dashes} ) {
+                infof("add new column 'dashes'");
+                $dbh->do(q{ALTER TABLE vrules ADD dashes VARCHAR(255) NOT NULL DEFAULT ''});
+            }
+            $dbh->commit;
+        }
+
+        # timestamp
+        {
+            $dbh->begin_work;
+            my $columns = $dbh->select_all(q{PRAGMA table_info("graphs")});
+            my %graphs_columns;
+            $graphs_columns{$_->{name}} = 1 for @$columns;
+            if ( ! exists $graphs_columns{timestamp} ) {
+                infof("add new column 'timestamp'");
+                $dbh->do(q{ALTER TABLE graphs ADD timestamp UNSIGNED INT DEFAULT NULL});
+            }
+            $dbh->commit;
+        }
+
         return;
     };
 }
@@ -166,38 +210,40 @@ sub get_by_id_for_rrdupdate_short {
     );
     return if !$data;
 
-    $dbh->begin_work;
-    my $subtract;
-    my $for_update = ( $dbh->connect_info->[0] =~ /^(?i:dbi):mysql:/ ) ? ' FOR UPDATE' : '';
-    my $prev = $dbh->select_row(
-        'SELECT * FROM prev_short_graphs WHERE graph_id = ?'.$for_update,
-        $data->{id}
-    );
-    if ( !$prev ) {
-        $subtract = 'U';
-        $dbh->query(
-            'INSERT INTO prev_short_graphs (graph_id, number, subtract, updated_at) 
-                         VALUES (?,?,?,?)',
-            $data->{id}, $data->{number}, undef, $data->{updated_at});
-    }
-    elsif ( $data->{updated_at} != $prev->{updated_at} ) {
-        $subtract = $data->{number} - $prev->{number};
-        $dbh->query(
-            'UPDATE prev_short_graphs SET number=?, subtract=?, updated_at=? WHERE graph_id = ?',
-            $data->{number}, $subtract, $data->{updated_at}, $data->{id}
+    unless ( $self->{disable_subtract} ) {
+        $dbh->begin_work;
+        my $subtract;
+        my $for_update = ( $dbh->connect_info->[0] =~ /^(?i:dbi):mysql:/ ) ? ' FOR UPDATE' : '';
+        my $prev = $dbh->select_row(
+            'SELECT * FROM prev_short_graphs WHERE graph_id = ?'.$for_update,
+            $data->{id}
         );
-    }
-    else {
-        if ( $data->{mode} eq 'gauge' || $data->{mode} eq 'modified' ) {
-            $subtract = $prev->{subtract};
-            $subtract = 'U' if ! defined $subtract;
+        if ( !$prev ) {
+            $subtract = 'U';
+            $dbh->query(
+                'INSERT INTO prev_short_graphs (graph_id, number, subtract, updated_at)
+                VALUES (?,?,?,?)',
+                $data->{id}, $data->{number}, undef, $data->{updated_at});
+        }
+        elsif ( $data->{updated_at} != $prev->{updated_at} ) {
+            $subtract = $data->{number} - $prev->{number};
+            $dbh->query(
+                'UPDATE prev_short_graphs SET number=?, subtract=?, updated_at=? WHERE graph_id = ?',
+                $data->{number}, $subtract, $data->{updated_at}, $data->{id}
+            );
         }
         else {
-            $subtract = 0;
+            if ( $data->{mode} eq 'gauge' || $data->{mode} eq 'modified' ) {
+                $subtract = $prev->{subtract};
+                $subtract = 'U' if ! defined $subtract;
+            }
+            else {
+                $subtract = 0;
+            }
         }
+        $dbh->commit;
+        $data->{subtract_short} = $subtract;
     }
-    $dbh->commit;
-    $data->{subtract_short} = $subtract;
     $self->inflate_row($data);
 }
 
@@ -211,46 +257,48 @@ sub get_by_id_for_rrdupdate {
     );
     return if !$data;
 
-    $dbh->begin_work;
-    my $subtract;
+    unless ( $self->{disable_subtract} ) {
+        $dbh->begin_work;
+        my $subtract;
 
-    my $for_update = ( $dbh->connect_info->[0] =~ /^(?i:dbi):mysql:/ ) ? ' FOR UPDATE' : '';
-    my $prev = $dbh->select_row(
-        'SELECT * FROM prev_graphs WHERE graph_id = ?' . $for_update,
-        $data->{id}
-    );
-    
-    if ( !$prev ) {
-        $subtract = 'U';
-        $dbh->query(
-            'INSERT INTO prev_graphs (graph_id, number, subtract, updated_at) 
-                         VALUES (?,?,?,?)',
-            $data->{id}, $data->{number}, undef, $data->{updated_at});
-    }
-    elsif ( $data->{updated_at} != $prev->{updated_at} ) {
-        $subtract = $data->{number} - $prev->{number};
-        $dbh->query(
-            'UPDATE prev_graphs SET number=?, subtract=?, updated_at=? WHERE graph_id = ?',
-            $data->{number}, $subtract, $data->{updated_at}, $data->{id}
-        );        
-    }
-    else {
-        if ( $data->{mode} eq 'gauge' || $data->{mode} eq 'modified' ) {
-            $subtract = $prev->{subtract};
-            $subtract = 'U' if ! defined $subtract;
+        my $for_update = ( $dbh->connect_info->[0] =~ /^(?i:dbi):mysql:/ ) ? ' FOR UPDATE' : '';
+        my $prev = $dbh->select_row(
+            'SELECT * FROM prev_graphs WHERE graph_id = ?' . $for_update,
+            $data->{id}
+        );
+
+        if ( !$prev ) {
+            $subtract = 'U';
+            $dbh->query(
+                'INSERT INTO prev_graphs (graph_id, number, subtract, updated_at)
+                VALUES (?,?,?,?)',
+                $data->{id}, $data->{number}, undef, $data->{updated_at});
+        }
+        elsif ( $data->{updated_at} != $prev->{updated_at} ) {
+            $subtract = $data->{number} - $prev->{number};
+            $dbh->query(
+                'UPDATE prev_graphs SET number=?, subtract=?, updated_at=? WHERE graph_id = ?',
+                $data->{number}, $subtract, $data->{updated_at}, $data->{id}
+            );
         }
         else {
-            $subtract = 0;
+            if ( $data->{mode} eq 'gauge' || $data->{mode} eq 'modified' ) {
+                $subtract = $prev->{subtract};
+                $subtract = 'U' if ! defined $subtract;
+            }
+            else {
+                $subtract = 0;
+            }
         }
-    }
 
-    $dbh->commit;
-    $data->{subtract} = $subtract;
+        $dbh->commit;
+        $data->{subtract} = $subtract;
+    }
     $self->inflate_row($data);
 }
 
 sub update {
-    my ($self, $service, $section, $graph, $number, $mode, $color ) = @_;
+    my ($self, $service, $section, $graph, $number, $mode, $color, $timestamp ) = @_;
     my $dbh = $self->dbh;
     $dbh->begin_work;
 
@@ -267,8 +315,8 @@ sub update {
         if ( $mode ne 'modified' || ($mode eq 'modified' && $data->{number} != $number) ) {
             $color ||= $data->{color};
             $dbh->query(
-                'UPDATE graphs SET number=?, mode=?, color=?, updated_at=? WHERE id = ?',
-                $number, $mode, $color, time, $data->{id}
+                'UPDATE graphs SET number=?, mode=?, color=?, updated_at=?, timestamp=? WHERE id = ?',
+                $number, $mode, $color, time, $timestamp, $data->{id}
             );
         }
     }
@@ -276,10 +324,10 @@ sub update {
         my @colors = List::Util::shuffle(qw/33 66 99 cc/);
         $color ||= '#' . join('', splice(@colors,0,3));
         $dbh->query(
-            'INSERT INTO graphs (service_name, section_name, graph_name, number, mode, color, llimit, sllimit, created_at, updated_at) 
-                         VALUES (?,?,?,?,?,?,?,?,?,?)',
-            $service, $section, $graph, $number, $mode, $color, -1000000000, -100000 ,time, time
-        ); 
+            'INSERT INTO graphs (service_name, section_name, graph_name, number, mode, color, llimit, sllimit, created_at, updated_at, timestamp)
+                         VALUES (?,?,?,?,?,?,?,?,?,?,?)',
+            $service, $section, $graph, $number, $mode, $color, -1000000000, -100000 , time, time, $timestamp
+        );
     }
 
     my $row = $self->dbh->select_row(
@@ -388,6 +436,28 @@ sub get_all_graph_all {
     \@ret;
 }
 
+sub get_all_graph_as_tree {
+    my ( $self )  = @_;
+    my $graphs = $self->get_all_graph_name();
+    my %services;
+    my @services;
+    for my $row ( @$graphs ) {
+        push @{$services{$row->{service_name}}->{$row->{section_name}}}, $row; 
+    }
+    for my $service_name ( sort { lc($a) cmp lc($b) } keys %services ) {
+        my @sections = map {{
+            name => $_,
+            graphs => $services{$service_name}->{$_}
+        }} sort { lc($a) cmp lc($b) } keys %{$services{$service_name}};
+        push @services, {
+            name => $service_name,
+            sections => \@sections
+        }
+    }
+    return \@services;
+}
+
+
 sub remove {
     my ($self, $id ) = @_;
     my $dbh = $self->dbh;
@@ -396,10 +466,12 @@ sub remove {
         'DELETE FROM graphs WHERE id = ?',
         $id
     );
-    $dbh->query(
-        'DELETE FROM prev_graphs WHERE graph_id = ?',
-        $id
-    );
+    unless ( $self->{disable_subtract} ) {
+        $dbh->query(
+            'DELETE FROM prev_graphs WHERE graph_id = ?',
+            $id
+        );
+    }
     $dbh->commit;
 
 }
@@ -515,6 +587,83 @@ sub get_all_complex_graph_all {
     return [] unless $list;
     my @ret = map { $self->inflate_complex_row($_) } @$list;
     \@ret;
+}
+
+sub update_vrule {
+    my ($self, $graph_path, $time, $color, $desc, $dashes) = @_;
+
+    $self->dbh->query(
+        'INSERT INTO vrules (graph_path,time,color,description,dashes) values (?,?,?,?,?)',
+        $graph_path, $time, $color, $desc, $dashes,
+    );
+
+   my $row = $self->dbh->select_row(
+        'SELECT * FROM vrules WHERE graph_path = ? AND time = ? AND color = ? AND description = ? AND dashes = ?',
+        $graph_path, $time, $color, $desc, $dashes,
+    );
+
+    return $row;
+}
+
+# "$span" is a parameter named "t",
+# "$from" and "$to" are paramters same nameed.
+sub get_vrule {
+    my ($self, $span, $from, $to, $graph_path) = @_;
+
+    my($from_time, $to_time) = (0, time);
+    # same rule as GrowthForecast::RRD#calc_period
+    if ( $span eq 'all' ) {
+        $from_time = 0;
+        $to_time   = 4294967295; # unsigned int max
+    } elsif ( $span eq 'c' || $span eq 'sc' ) {
+        if ($from =~ /\A[1-9][0-9]*\z/) {
+            $from_time = $from;
+        } else {
+            $from_time = HTTP::Date::str2time($from);
+            die "invalid from date: $from" unless $from_time;
+        }
+        if ($to && $to =~ /\A[1-9][0-9]*\z/) {
+            $to_time = $to;
+        } else {
+            $to_time = $to ? HTTP::Date::str2time($to) : time;
+            die "invalid to date: $to" unless $to_time;
+        }
+        die "from($from) is newer than to($to)" if $from_time > $to_time;
+    } elsif ( $span eq 'h' || $span eq 'sh' ) {
+        $from_time = time -1 * 60 * 60 * 2;
+    } elsif ( $span eq 'n' || $span eq 'sn' ) {
+        $from_time = time -1 * 60 * 60 * 14;
+    } elsif ( $span eq 'w' ) {
+        $from_time = time -1 * 60 * 60 * 24 * 8;
+    } elsif ( $span eq 'm' ) {
+        $from_time = time -1 * 60 * 60 * 24 * 35;
+    } elsif ( $span eq 'y' ) {
+        $from_time = time -1 * 60 * 60 * 24 * 400;
+    } elsif ( $span eq '3d' ) {
+        $from_time = time -1 * 60 * 60 * 24 * 3;
+    } elsif ( $span eq '8h' ) {
+        $from_time = time -1 * 8 * 60 * 60;
+    } elsif ( $span eq '4h' ) {
+        $from_time = time -1 * 4 * 60 * 60;
+    } else {
+        $from_time = time -1 * 60 * 60 * 33; # 33 hours
+    }
+
+    my @vrules = ();
+
+    my @gp = split '/', substr($graph_path, 1), 3;
+    my $ph = ',?'x@gp;
+    my @path;
+    for (my $i=0; $i<@gp; $i++ ) {
+        push @path, "/".join("/",@gp[0..$i]); 
+    }
+    my $rows = $self->dbh->select_all(
+        'SELECT * FROM vrules WHERE (time BETWEEN ? and ?) AND graph_path in ("/"'.$ph.')',
+        $from_time, $to_time, @path
+    );
+    push @vrules, @$rows;
+
+    return @vrules;
 }
 
 1;
